Split Redis component into BackendRedis and SystemRedis components
