+++ pkg/3scale/amp/template/amp.go
@@ -25,7 +25,8 @@ func (f *AmpTemplateFactory) Adapters() []adapters.Adapter {
 	return []adapters.Adapter{
 		adapters.NewImagesAdapter(),
 		adapters.NewSystemMysqlImageAdapter(),
-		adapters.NewRedisAdapter(),
+		adapters.NewBackendRedisAdapter(),
+		adapters.NewSystemRedisAdapter(),
 		adapters.NewBackendAdapter(false),
 		adapters.NewMysqlAdapter(),
 		adapters.NewMemcachedAdapter(),
